+++ src/globals.h
@@ -58,7 +58,7 @@ public:
 	static int FromChunkX, FromChunkZ, ToChunkX, ToChunkZ; // Current area of world being rendered
 	static size_t MapsizeZ, MapsizeX, Terrainsize; // size of that area in blocks (no offset)
 	static int MapminY; //minimum height to render from
-	static unsigned int MapsizeY; //maximum height to render (max CHUNKSIZE_Y (256))
+	static size_t MapsizeY; //maximum height to render (max CHUNKSIZE_Y (256))
 	static int OffsetY; // y pixel offset in the final image for one y step in 3d array (2)
 	static Settings settings; //Used settings
